pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "city-automation"
version = "1.0.0"
description = "Framework d'automatisation pour BlueStacks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Xavier"}
]

# Note: PyTorch doit être installé séparément avec la bonne version (CPU ou CUDA)
# CPU: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# GPU: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

dependencies = [
    "mss>=9.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pyautogui>=0.9.54",
    "pynput>=1.7.6",
    "pywin32>=306; sys_platform == 'win32'",
    "easyocr>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.8.0",
    "pyright>=1.1.0",
]
ocr-tesseract = [
    "pytesseract>=0.3.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["manoirs", "core", "actions", "vision", "utils"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

[tool.setuptools.packages.find]
include = ["manoirs*", "core*", "actions*", "vision*", "utils*", "erreurs*"]

# ============================================
# Ruff - Linting et formatage
# ============================================
[tool.ruff]
line-length = 100
target-version = "py39"
exclude = [
    ".git",
    "__pycache__",
    "data",
    "logs",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort (tri des imports)
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade (modernisation Python)
    "B",      # bugbear (bugs courants)
    "SIM",    # simplify (simplifications)
]
ignore = [
    "E501",   # ligne trop longue (gere par formatter)
    "B027",   # empty method in ABC without @abstractmethod (voulu)
    "SIM102", # nested if (parfois plus lisible)
    "SIM108", # ternary operator (parfois moins lisible)
    "SIM117", # multiple with statements (parfois plus lisible)
    "B904",   # raise from (pas toujours necessaire)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]  # imports not at top (necessaire pour path setup)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================
# Pyright - Type checking
# ============================================
[tool.pyright]
include = ["core", "actions", "vision", "utils", "erreurs", "manoirs"]
exclude = ["tests", "data", "logs"]
pythonVersion = "3.9"
typeCheckingMode = "basic"  # "basic" pour commencer, "strict" plus tard
